requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
numpy==1.16.6
pandas==0.24.2
pycrypto==2.6.1
pygobject==3.26.1
python-dateutil==2.8.2
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
scikit-learn==0.20.4
scipy==1.2.3
SecretStorage==2.3.1
six==1.16.0
sklearn==0.0
